content/base/test/test_CSP_inlinestyle.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Content Security Policy inline stylesheets stuff</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<iframe style="width:100%;height:300px;" id='cspframe'></iframe>
<iframe style="width:100%;height:300px;" id='cspframe2'></iframe>
<iframe style="width:100%;height:300px;" id='cspframe3'></iframe>
<script class="testbody" type="text/javascript">

var path = "/tests/content/base/test/";


//////////////////////////////////////////////////////////////////////
// set up and go
SimpleTest.waitForExplicitFinish();

var done = 0;

// Our original CSP implementation does not block inline styles.
function checkStyles(evt) {
  var cspframe = document.getElementById('cspframe');
  var color;

  // black means the style wasn't applied.  green colors are used for styles
  //expected to be applied.  A color is red if a style is erroneously applied
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('linkstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'External Stylesheet (original CSP implementation) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('inlinestylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Inline Style TAG (original CSP implementation) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('attrstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Style Attribute (original CSP implementation) (' + color + ')');
  // SMIL tests
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('xmlTest',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'XML Attribute styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('cssOverrideTest',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Override styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('cssOverrideTestById',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Override styling via ID lookup (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe.contentDocument.getElementById('cssSetTestById',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Set Element styling via ID lookup (SMIL) (' + color + ')');
  checkIfDone();
}

// When a CSP 1.0 compliant policy is specified we should block inline
// styles applied by <style> element, style attribute, and SMIL <animate> and <set> tags
// (when it's not explicitly allowed.)
function checkStylesSpecCompliant(evt) {
  var cspframe = document.getElementById('cspframe2');
  var color;

  // black means the style wasn't applied.  green colors are used for styles
  //expected to be applied.  A color is red if a style is erroneously applied
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('linkstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'External Stylesheet (CSP 1.0 spec compliant) (' + color + ')');
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('inlinestylediv'),null)['color'];
  ok('rgb(0, 0, 0)' === color, 'Inline Style TAG (CSP 1.0 spec compliant) (' + color + ')');
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('attrstylediv'),null)['color'];
  ok('rgb(0, 0, 0)' === color, 'Style Attribute (CSP 1.0 spec compliant) (' + color + ')');
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('csstextstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'cssText (CSP 1.0 spec compliant) (' + color + ')');
  // SMIL tests
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('xmlTest',null))['fill'];
  ok('rgb(0, 0, 0)' === color, 'XML Attribute styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('cssOverrideTest',null))['fill'];
  ok('rgb(0, 0, 0)' === color, 'CSS Override styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('cssOverrideTestById',null))['fill'];
  ok('rgb(0, 0, 0)' === color, 'CSS Override styling via ID lookup (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('cssSetTestById',null))['fill'];
  ok('rgb(0, 0, 0)' === color, 'CSS Set Element styling via ID lookup (SMIL) (' + color + ')');

  color = window.getComputedStyle(cspframe2.contentDocument.getElementById('modifycsstextdiv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Modify loaded style sheet via cssText (' + color + ')');

  checkIfDone();
}

// When a CSP 1.0 compliant policy is specified we should allow inline
// styles when it is explicitly allowed.
function checkStylesSpecCompliantAllowed(evt) {
  var cspframe = document.getElementById('cspframe3');
  var color;

  // black means the style wasn't applied.  green colors are used for styles
  // expected to be applied.  A color is red if a style is erroneously applied
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('linkstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'External Stylesheet (CSP 1.0 spec compliant, allowed) (' + color + ')');
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('inlinestylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Inline Style TAG (CSP 1.0 spec compliant, allowed) (' + color + ')');
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('attrstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Style Attribute (CSP 1.0 spec compliant, allowed) (' + color + ')');

  // Note that the below test will fail if "script-src: 'unsafe-inline'" breaks,
  // since it relies on executing script to set .cssText
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('csstextstylediv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'style.cssText (CSP 1.0 spec compliant, allowed) (' + color + ')');
  // SMIL tests
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('xmlTest',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'XML Attribute styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('cssOverrideTest',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Override styling (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('cssOverrideTestById',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Override styling via ID lookup (SMIL) (' + color + ')');
  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('cssSetTestById',null))['fill'];
  ok('rgb(0, 255, 0)' === color, 'CSS Set Element styling via ID lookup (SMIL) (' + color + ')');

  color = window.getComputedStyle(cspframe3.contentDocument.getElementById('modifycsstextdiv'),null)['color'];
  ok('rgb(0, 255, 0)' === color, 'Modify loaded style sheet via cssText (' + color + ')');

  checkIfDone();
}

function checkIfDone() {
  done++;
  if (done == 3)
    SimpleTest.finish();
}

SpecialPowers.pushPrefEnv(
  {'set':[["security.csp.speccompliant", true]]},
  function() {
    // save this for last so that our listeners are registered.
    // ... this loads the testbed of good and bad requests.
    document.getElementById('cspframe').src = 'file_CSP_inlinestyle_main.html';
    document.getElementById('cspframe').addEventListener('load', checkStyles, false);
    document.getElementById('cspframe2').src = 'file_CSP_inlinestyle_main_spec_compliant.html';
    document.getElementById('cspframe2').addEventListener('load', checkStylesSpecCompliant, false);
    document.getElementById('cspframe3').src = 'file_CSP_inlinestyle_main_spec_compliant_allowed.html';
    document.getElementById('cspframe3').addEventListener('load', checkStylesSpecCompliantAllowed, false);
  }
);
</script>
</pre>
</body>
</html>
